component restructured
