+++ src/Components/About/About.tsx
@@ -1,7 +1,6 @@
 const About = () => {
   return (
     <div className="mt-96 md:mt-48 lg:mt-0">
-      <div className="text-center text-4xl py-5">All about Cloud Community days</div>
       <div className="grid sm:grid-cols-1 md:grid-cols-2  place-items-center bg-lightBlue bg-opacity-50">
         <div className="w-full lg:w-3/4 my-0 mx-auto pl-20">
           <h3 className="text-gray-800 text-3xl py-5">What is CCD’22 ?</h3>
